Switch from linenoise to BSD editline
linenoise had shortcomings which made it really annoying to use it some
edge cases. The biggest one being that upstream antirez/linenoise is not
properly mentioned. Some essential features, such as unicode support,
even required using a fork.

I am not perfectly happy with libedit either, most importantly I am
quite annoyed by the fact that it doesn't provide a completion API, but
it seems to be the lesser of two evils.

While at it:

* Don't compile with -Wconversion because that becomes very annoying
  when computing pointer differences and using them as sizes.
* Make sure that the cleanup function isn't invoked more than once to
  prevent a double-free.