complete: use editline's wide character API
The underlying problem is that el_deletestr always deletes characters,
not bytes even when using the multibyte string API. Previously this
issue was solved by using both el_wline and el_line and calculating
character and byte cursor positions in the complete() function.

This commit changes this and always uses the wide char API and character
cursor to solve the underlying problem in a much cleaner way.