input: block all signals in atexit(3) handler
This fixes a race condition I occasionally witnessed on OpenBSD where a
signal would be delivered to the input process. This signal would cause
el_gets(3) to fail with EINTR, thereby causing input to call err(3)
which would invoke cleanup() at exit. During the invocation of cleanup()
a signal would be delivered to the process causing it to jump to the
signal handler before completing the cleanup operation. The result would
often be a corrupted history file.