Switch from editline to GNU readline
I encountered quite a few bugs with editline in the past few months.
Additionally, editline bugs vary between different BSD versions. For
this reason, I decided to switch to GNU readline. The readline functions
used are compatible with the readline compatibility layer provided by
many editline implementations. Therefore, it still possible to use this
software with BSD editline.

In the long run it might also be desirable to support more minimal
implementations of a readline compatibility layer, e.g.
<https://github.com/troglobit/editline>, to cover use case where
ncurses is not desired.