Setup cleanup handlers after initialization finished
Otherwise the history file might get truncated if a signal is received
during the initialization because the history might not be loaded yet
but is still written to the history file.

This change has the disadvantage that temporary file may be left laying
around if a terminating signal is received after finishing
initialization but before setting up the signal handlers. I consider
this to be less worse than truncating the history though.